Add files via upload
Method 1 : With 2 separate strings to store the output, separate for loops to store the encoded and decoded message and separate if loop to check the direction before calling the function. Lot of code repetition
Method 2: One string result to store the end string. This again is repetitive.
Method 3: The code is pretty same for both encrypt and decrypt except the shift_amount. So what we do is keep an if and shift if it is decode in which case we modift the shift to include a minus 1 otherwise the function is reduced to one for loop
Best solution